Add groups_assignment for IDPs
Fix assignments
