+++ examples/okta_idp_social/basic.tf
@@ -44,4 +44,11 @@ resource okta_idp_social microsoft {
   client_id         = "abcd123"
   client_secret     = "abcd123"
   username_template = "idpuser.userPrincipalName"
+  groups_action     = "ASSIGN"
+  groups_assignment = ["${okta_group.test.id}"]
+}
+
+resource "okta_group" "test" {
+  name        = "testAcc_replace_with_uuid"
+  description = "testing, testing, testing"
 }
