+++ okta/idp_models.go
@@ -60,6 +60,7 @@ type (
 
 	IDPGroupsAction struct {
 		Action              string   `json:"action,omitempty"`
+		Assignments         []string `json:"assignments,omitempty"`
 		Filter              []string `json:"filter,omitempty"`
 		SourceAttributeName string   `json:"sourceAttributeName,omitempty"`
 	}
